requirements.txt
aiogram==3.3.0
loguru==0.7.2
python-dotenv==1.0.1
Telethon==1.35.0